refactor: update StakingInfo hook
